Refactor KeystoneContext definition for docs
