+++ .changeset/perfect-pillows-prove.md
@@ -1,0 +1,5 @@
+---
+'@keystone-next/types': patch
+---
+
+Refactored `KeystoneContext` definition to make documentation easier. No functional changes.
