+++ packages-next/types/src/index.ts
@@ -4,3 +4,4 @@ export * from './utils';
 export * from './session';
 export * from './admin-meta';
 export * from './base';
+export * from './context';
